# Speed up System.Drawing.Color factory and HSB/HSL methods