--- a/src/System.Drawing.Primitives/System.Drawing.Primitives.sln
+++ b/src/System.Drawing.Primitives/System.Drawing.Primitives.sln
@@ -7,6 +7,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Drawing.Primitives.T
 		{8F472B93-574C-4AEC-9D28-6C2360A55BBF} = {8F472B93-574C-4AEC-9D28-6C2360A55BBF}
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Drawing.Primitives.Performance.Tests", "tests\Performance\System.Drawing.Primitives.Performance.Tests.csproj", "{1BD5C9BF-D7F2-4249-AA31-43B1850A5DB3}"
+	ProjectSection(ProjectDependencies) = postProject
+		{8F472B93-574C-4AEC-9D28-6C2360A55BBF} = {8F472B93-574C-4AEC-9D28-6C2360A55BBF}
+	EndProjectSection
+EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Drawing.Primitives", "src\System.Drawing.Primitives.csproj", "{8F472B93-574C-4AEC-9D28-6C2360A55BBF}"
 	ProjectSection(ProjectDependencies) = postProject
 		{10F74537-6423-48F5-A7F3-4DE94E42AF8F} = {10F74537-6423-48F5-A7F3-4DE94E42AF8F}
@@ -30,6 +35,10 @@ Global
 		{CF54638C-A382-4A78-9AD6-2304CEEFEB01}.Debug|Any CPU.Build.0 = netcoreapp-Debug|Any CPU
 		{CF54638C-A382-4A78-9AD6-2304CEEFEB01}.Release|Any CPU.ActiveCfg = netcoreapp-Release|Any CPU
 		{CF54638C-A382-4A78-9AD6-2304CEEFEB01}.Release|Any CPU.Build.0 = netcoreapp-Release|Any CPU
+		{1BD5C9BF-D7F2-4249-AA31-43B1850A5DB3}.Debug|Any CPU.ActiveCfg = netcoreapp-Debug|Any CPU
+		{1BD5C9BF-D7F2-4249-AA31-43B1850A5DB3}.Debug|Any CPU.Build.0 = netcoreapp-Debug|Any CPU
+		{1BD5C9BF-D7F2-4249-AA31-43B1850A5DB3}.Release|Any CPU.ActiveCfg = netcoreapp-Release|Any CPU
+		{1BD5C9BF-D7F2-4249-AA31-43B1850A5DB3}.Release|Any CPU.Build.0 = netcoreapp-Release|Any CPU
 		{8F472B93-574C-4AEC-9D28-6C2360A55BBF}.Debug|Any CPU.ActiveCfg = netcoreapp-Windows_NT-Debug|Any CPU
 		{8F472B93-574C-4AEC-9D28-6C2360A55BBF}.Debug|Any CPU.Build.0 = netcoreapp-Windows_NT-Debug|Any CPU
 		{8F472B93-574C-4AEC-9D28-6C2360A55BBF}.Release|Any CPU.ActiveCfg = netcoreapp-Windows_NT-Release|Any CPU
@@ -44,6 +53,7 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{CF54638C-A382-4A78-9AD6-2304CEEFEB01} = {1A2F9F4A-A032-433E-B914-ADD5992BB178}
+		{1BD5C9BF-D7F2-4249-AA31-43B1850A5DB3} = {1A2F9F4A-A032-433E-B914-ADD5992BB178}
 		{8F472B93-574C-4AEC-9D28-6C2360A55BBF} = {E107E9C1-E893-4E87-987E-04EF0DCEAEFD}
 		{10F74537-6423-48F5-A7F3-4DE94E42AF8F} = {2E666815-2EDB-464B-9DF6-380BF4789AD4}
 	EndGlobalSection