As evidenced by https://github.com/dotnet/corefx/issues/31820, some apps make a huge number of calls to the `Color` methods.  This change optimizes the `Color` factory and HSB/HSL methods to make those apps somewhat less slow.

``` ini

BenchmarkDotNet=v0.11.0, OS=Windows 10.0.17134.228 (1803/April2018Update/Redstone4)
Intel Xeon CPU E3-1505M v6 3.00GHz, 1 CPU, 8 logical and 4 physical cores
Frequency=2929686 Hz, Resolution=341.3335 ns, Timer=TSC
.NET Core SDK=2.1.400
  [Host] : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  before : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  after  : .NET Core ? (CoreCLR 4.6.26816.0, CoreFX 4.6.26816.0), 64bit RyuJIT

Jit=RyuJit  Platform=X64  Toolchain=Default  

```
|            Method |    Job |       Mean |     Error |    StdDev | Scaled |
|------------------ |------- |-----------:|----------:|----------:|-------:|
| FromArgb(A,Color) | before | 3,413.7 ns | 26.012 ns | 24.332 ns |   1.00 |
| FromArgb(A,Color) |  after | 1,047.6 ns |  7.910 ns |  7.399 ns |   0.31 |
|                   |        |            |           |           |        |
| FromArgb(A,R,G,B) | before | 2,974.0 ns | 22.484 ns | 21.031 ns |   1.00 |
| FromArgb(A,R,G,B) |  after |   797.6 ns |  5.037 ns |  4.465 ns |   0.27 |
|                   |        |            |           |           |        |
|    FromArgb(ARGB) | before |   316.5 ns |  2.405 ns |  2.250 ns |   1.00 |
|    FromArgb(ARGB) |  after |   316.6 ns |  2.325 ns |  2.175 ns |   1.00 |
|                   |        |            |           |           |        |
|     GetBrightness | before | 3,011.9 ns | 26.623 ns | 20.786 ns |   1.00 |
|     GetBrightness |  after |   975.6 ns | 19.204 ns | 21.345 ns |   0.32 |
|                   |        |            |           |           |        |
|            GetHue | before | 5,585.2 ns | 23.481 ns | 20.815 ns |   1.00 |
|            GetHue |  after | 1,288.7 ns | 15.075 ns | 14.101 ns |   0.23 |
|                   |        |            |           |           |        |
|     GetSaturation | before | 2,794.9 ns | 28.432 ns | 25.205 ns |   1.00 |
|     GetSaturation |  after | 1,198.2 ns | 17.657 ns | 15.652 ns |   0.43 |

The HSB methods have slightly different behavior in that they keep the values in fixed-point for as long as possible.  The old logic converted to float immediately, which results in slightly lower precision.  For example, `Color.DarkOliveGreen.GetHue()` returns 81.99999 under the old implementation but returns the more correct value of 82 in the new.  `Color.FromArgb(255, 255, 18).GetBrightness() == Color.FromArgb(137, 137, 136).GetBrightness()` evaluates to false in the old implementation but is (correctly) true in the new.  No differences are greater than that caused by the increased precision.

Times are based on [this](https://gist.github.com/saucecontrol/e8911e8db288a1b74897ef976b2c8c33) benchmark code.