@AraHaan 2018-08-19
Thanks, however I did not notice it could be optimized further. The only slowness I find though is when using SetPixel() on a image with like a 5,000 x 3,000 or so pixel size (or in this case images larger than 800 x 800).