@saucecontrol 2018-08-19
>> Thanks, however I did not notice it could be optimized further.

The slowness in the `Color` methods will be dwarfed by the slowness of `GetPixel()` and `SetPixel()`.  In reality, `Color` is a terrible way to represent a pixel value.  This PR just makes it slightly less terrible when used in that way.