@safern 2018-08-20
Could you also make sure there is enough test coverage for this members (if not add tests to not cause regressions) and I think it might be worth checking in perf tests.