@saucecontrol 2018-08-20
Test coverage is quite good already.  The only hole I see is that `GetSaturation()` could use some combinations that return values other than 0 and 1.  I'm not seeing any perf tests for S.D.Primitives.  Is there a guide for those?