@safern 2018-08-20
> The only hole I see is that GetSaturation() could use some combinations that return values other than 0 and 1.

Would be great to add them. 

> I'm not seeing any perf tests for S.D.Primitives. Is there a guide for those?

I would just say copy System.Drawing.Common/tests/Performance structure or any other perf tests that you'd like. Then just follow: https://github.com/dotnet/corefx/blob/653d41fb63152ed920b516b3706cd42b1275b474/Documentation/project-docs/performance-tests.md