@saucecontrol 2018-08-21
Thanks.  Added a perf project and a couple of test cases for `GetSaturation()`