@saucecontrol 2018-08-21
Tidied up a couple of things and updated the description with the final benchmark numbers.